--- bin/cache
@@ -1,43 +1,0 @@
-#!/usr/bin/env ruby
-require "fileutils"
-require_relative "./helpers/helpers.rb"
-
-# path to your application root.
-APP_ROOT = File.expand_path("..", __dir__)
-
-# `exist?` is defined in new ruby. Monkey patch for old versions
-if File.methods.include?(:exists?) && !File.methods.include?(:exist?)
-  class File
-    def self.exist? file_name
-      self.exists? file_name
-    end
-  end
-end
-
-puts "Toggle App Cache in development mode"
-cache_toggle_file = "tmp/caching-dev.txt"
-
-FileUtils.chdir APP_ROOT do
-  if File.exist?(cache_toggle_file)
-    puts "File '#{cache_toggle_file}' exists"
-    FileUtils.rm_rf(cache_toggle_file, verbose: true)
-    puts "File '#{cache_toggle_file}' is removed"
-    puts "Cache is OFF"
-  else
-    puts "File '#{cache_toggle_file}' does not exist"
-    FileUtils.touch(cache_toggle_file, verbose: true)
-    puts "File '#{cache_toggle_file}' is created"
-    puts "Cache is ON"
-  end
-end
-
-# Not sure if for windows it will work.
-# https://stackoverflow.com/questions/11982057/how-can-i-trigger-a-shell-script-and-run-in-background-async-in-ruby
-#
-puts "Restarting PUMA server"
-if inside_rails_conainer?
-  system(@puma_stop_command)
-  system (@puma_start_command + " & ")
-else
-  puma_restart
-end
