--- bin/helpers/chewy.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-def chewy_index
-  puts 'Indexing with CHEWY'
-  container_exec('rails', 'rake chewy:update')
-end
-
-def chewy_reset
-  puts 'Dropping CHEWY indicies'
-  container_exec('rails', 'rake chewy:reset')
-end
