--- bin/helpers/cron.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-def cron_start
-  container_exec('--user root rails', '/etc/init.d/cron start')
-end
-
-def cron_stop
-  container_exec('--user root rails', '/etc/init.d/cron stop')
-end
