--- bin/index
@@ -1,9 +1,0 @@
-#!/usr/bin/env ruby
-require_relative "./helpers/helpers.rb"
-
-start_all_containers
-
-rails_install_gems
-rails_db_migrate
-
-chewy_index
