+++ bin/rails7startkit/helpers.rb
@@ -1,23 +1,10 @@
 # frozen_string_literal: true
 
-require_relative './common'
-require_relative './docker'
-require_relative './chewy'
-require_relative './sidekiq'
-require_relative './cron'
-require_relative './whenever'
-require_relative './rails'
-require_relative './puma'
-require_relative './rubocop'
-
 DELAY = 3
 ENV_NAME = ENV.fetch('RAILS_ENV', 'development')
 
 $steps_counter = 1
 $steps_messages = []
-
-@puma_start_command = 'bundle exec puma -C config/_PUMA.rb'
-@puma_stop_command  = 'pkill -f puma'
 
 def wait(details, delay = DELAY)
   puts "Waiting #{delay} seconds #{details}"
