--- bin/reset
@@ -1,23 +1,0 @@
-#!/usr/bin/env ruby
-require "fileutils"
-require_relative "./helpers/helpers.rb"
-
-# path to your application root.
-APP_ROOT = File.expand_path("..", __dir__)
-
-puts "This command clean up ALL THE DATA for the project!"
-prompt!
-
-FileUtils.chdir APP_ROOT do
-  FileUtils.rm_rf('db/PGSQL/', verbose: true)
-  FileUtils.rm_rf('db/REDIS/', verbose: true)
-  FileUtils.rm_rf('db/ELASTIC/', verbose: true)
-
-  FileUtils.mkdir_p('db/PGSQL/', verbose: true)
-  FileUtils.mkdir_p('db/REDIS/', verbose: true)
-  FileUtils.mkdir_p('db/ELASTIC/', verbose: true)
-
-  FileUtils.touch('db/PGSQL/.keep', verbose: true)
-  FileUtils.touch('db/REDIS/.keep', verbose: true)
-  FileUtils.touch('db/ELASTIC/.keep', verbose: true)
-end
