--- bin/restart
@@ -1,3 +1,0 @@
-#!/usr/bin/env ruby
-system('bin/stop')
-system('bin/start')
