--- bin/restart_all
@@ -1,3 +1,0 @@
-#!/usr/bin/env ruby
-system('bin/stop_all')
-system('bin/start_all')
