--- bin/start
@@ -1,11 +1,0 @@
-#!/usr/bin/env ruby
-require_relative "./helpers/helpers.rb"
-
-rails_install_gems
-rails_db_migrate
-
-cron_start
-chewy_index
-
-sidekiq_start
-puma_start
