--- bin/start_all
@@ -1,6 +1,0 @@
-#!/usr/bin/env ruby
-require_relative "./helpers/helpers.rb"
-
-start_all_containers
-system('bin/start')
-containers_information
