--- bin/status
@@ -1,6 +1,0 @@
-#!/usr/bin/env ruby
-require_relative "./helpers/helpers.rb"
-containers_information
-
-container_bash_exec('rails', 'ps a | grep puma')
-container_bash_exec('rails', 'ps a | grep sidekiq')
