--- bin/stop_all
@@ -1,4 +1,0 @@
-#!/usr/bin/env ruby
-require_relative "./helpers/helpers.rb"
-stop_all_containers
-containers_information
